23.4.1-38

+++ rpms-nws/Installer.HeatRiskIndex/component.spec
@@ -16,9 +16,18 @@ Packager: %{_build_site}
 Provides: awips2-heatriskindex
 Requires: awips2
 Requires: awips2-apps
+Requires: awips2-heatriskindex-data
 
 %description
 AWIPS II Heat Risk Index - Installs AWIPS II HeatRiskIndex Application.
+
+%package data
+Summary: Installation of static data used in awips2-heatriskindex
+Group: AWIPSII
+BuildArch: noarch
+
+%description data
+AWIPS II Heat Risk Index - Installs static data used in AWIPS II HeatRiskIndex Application.
 
 %prep
 # Verify That The User Has Specified A BuildRoot.
@@ -36,31 +45,27 @@ mkdir --parents %{_build_root}
 %install
 APP_DIR="%{_build_root}/awips2/apps/HeatRiskIndex"
 mkdir --parents ${APP_DIR}
-mkdir --parents ${APP_DIR}/data
-mkdir --parents ${APP_DIR}/install
-mkdir --parents ${APP_DIR}/runtime
-mkdir --parent ${APP_DIR}/install/PrismHiRes/logs
-mkdir --parent ${APP_DIR}/install/HeatRisk/logs
+
 SRC_DIR="%{_baseline_workspace}/apps/HeatRiskIndex"
 rsync --archive ${SRC_DIR}/ ${APP_DIR}/
 
 COMMON_DIR="%{_build_root}/awips2/edex/data/share/HeatRiskIndex"
 mkdir --parents ${COMMON_DIR}
-mkdir --parents ${COMMON_DIR}/data
-mkdir --parents ${COMMON_DIR}/data/stage
-mkdir --parents ${COMMON_DIR}/data/stage/PrismHiRes
-mkdir --parents ${COMMON_DIR}/data/stage/HeatRisk
-mkdir --parents ${COMMON_DIR}/data/climo
-mkdir --parents ${COMMON_DIR}/data/climo/PrismHiRes
-mkdir --parents ${COMMON_DIR}/data/climo/HeatRisk
-mkdir --parents ${COMMON_DIR}/data/runtime
-mkdir --parents ${COMMON_DIR}/data/runtime/PrismHiRes
-mkdir --parents ${COMMON_DIR}/data/runtime/HeatRisk
 mkdir --parents ${COMMON_DIR}/scripts
-mkdir --parents ${COMMON_DIR}/scripts/PrismHiRes
-mkdir --parents ${COMMON_DIR}/scripts/HeatRisk
-cp ${APP_DIR}/install/PrismHiRes/install/makeYearClimoTemps.sh ${COMMON_DIR}/scripts/PrismHiRes/.
-cp ${APP_DIR}/install/HeatRisk/install/installGrids.sh ${COMMON_DIR}/scripts/HeatRisk/.
+cp ${APP_DIR}/bin/LoadHeatRisk.sh ${COMMON_DIR}/scripts/.
+rm --recursive --force ${APP_DIR}/bin
+
+mkdir -p ${COMMON_DIR}/data
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+DATA_SRC_DIR="%{_static_files}/heatrisk"
+
+cp -r ${DATA_SRC_DIR}/heatrisk.hdf5 ${COMMON_DIR}/data/.
+if [ $? -ne 0 ]; then
+   exit 1
+fi
 
 %clean
 rm --recursive --force %{_build_root}
@@ -68,56 +73,25 @@ rm --recursive --force %{_build_root}
 %files
 %defattr(644,awips,fxalpha,755)
 %dir /awips2/apps/HeatRiskIndex
-%dir /awips2/apps/HeatRiskIndex/install
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/bin
-/awips2/apps/HeatRiskIndex/install/HeatRisk/bin/*
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/config
-/awips2/apps/HeatRiskIndex/install/HeatRisk/config/*
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/data
-/awips2/apps/HeatRiskIndex/install/HeatRisk/data/*
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/etc
-/awips2/apps/HeatRiskIndex/install/HeatRisk/etc/*
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/gfe
-/awips2/apps/HeatRiskIndex/install/HeatRisk/gfe/*
-%dir /awips2/apps/HeatRiskIndex/install/HeatRisk/install
-%dir /awips2/apps/HeatRiskIndex/install/PrismHiRes
-%dir /awips2/apps/HeatRiskIndex/install/PrismHiRes/bin
-/awips2/apps/HeatRiskIndex/install/PrismHiRes/bin/*
-%dir /awips2/apps/HeatRiskIndex/install/PrismHiRes/etc
-/awips2/apps/HeatRiskIndex/install/PrismHiRes/etc/*
-%dir /awips2/apps/HeatRiskIndex/install/PrismHiRes/gfe
-/awips2/apps/HeatRiskIndex/install/PrismHiRes/gfe/*
-%dir /awips2/apps/HeatRiskIndex/install/PrismHiRes/install
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/HeatRisk/install/install.sh
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/HeatRisk/install/installGrids.sh
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/HeatRisk/install/stageData_HeatRisk.sh
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/PrismHiRes/install/install.sh
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/PrismHiRes/install/stageData_PrismHiRes.sh
-%attr(755,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/PrismHiRes/install/makeYearClimoTemps.sh
-%dir /awips2/apps/HeatRiskIndex/runtime
+%dir /awips2/apps/HeatRiskIndex/config
+/awips2/apps/HeatRiskIndex/config/*
+%dir /awips2/apps/HeatRiskIndex/etc
+/awips2/apps/HeatRiskIndex/etc/*
+%dir /awips2/edex/data/share/HeatRiskIndex/data
 
 %attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/stage
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/stage/PrismHiRes
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/stage/HeatRisk
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/climo
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/climo/PrismHiRes
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/climo/HeatRisk
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/runtime
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/runtime/PrismHiRes
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/data/runtime/HeatRisk
 %attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/scripts
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/scripts/PrismHiRes
-%attr(775,awips,fxalpha) %dir /awips2/edex/data/share/HeatRiskIndex/scripts/HeatRisk
-%attr(775,awips,fxalpha) /awips2/edex/data/share/HeatRiskIndex/scripts/PrismHiRes/makeYearClimoTemps.sh
-%attr(775,awips,fxalpha) /awips2/edex/data/share/HeatRiskIndex/scripts/HeatRisk/installGrids.sh
-%attr(775,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/PrismHiRes/bin/LoadClimateGrids.sh
-%attr(775,awips,fxalpha) /awips2/apps/HeatRiskIndex/install/HeatRisk/bin/LoadHeatRisk.sh
+%attr(775,awips,fxalpha) /awips2/edex/data/share/HeatRiskIndex/scripts/LoadHeatRisk.sh
 %defattr(644,root,root,-)
 
+%files data
+%defattr(775,awips,fxalpha,775)
+/awips2/edex/data/share/HeatRiskIndex/data/*
+
 %changelog
+* Wed Feb 28 2024 Michael Gamazaychikov <michael.gamazaychikov@noaa.gov>
+- Removed old structure, added data subpackage
+
 * Tue Oct 17 2023 Michael Gamazaychikov <michael.gamazaychikov@noaa.gov>
 - Added share directories
